Extract user creation to users controller
